my_timer using A2 timer

+++ test/test.c
@@ -28,11 +28,11 @@ static void nand_read_write_perf(void *pvParameters)
     printf("test done\n");
 
     volatile double read_latency = latency_read(100);
-    printf("read done, latency: %d micro seconds\n", (uint16_t)(read_latency*1000000));
+    printf("read done, latency: %d micro seconds\n", (uint32_t)(read_latency*1000000));
     volatile double write_latency = latency_write(100);
-    printf("write done, latency: %d micro seconds\n", (uint16_t)(write_latency*1000000));
+    printf("write done, latency: %d micro seconds\n", (uint32_t)(write_latency*1000000));
     volatile double erase_latency = latency_erase(10);
-    printf("erase done, latency: %d micro seconds\n", (uint16_t)(erase_latency*1000000));
+    printf("erase done, latency: %d micro seconds\n", (uint32_t)(erase_latency*1000000));
 
 #endif
